+++ CHANGELOG.md
@@ -7,11 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.2] - 2020-11-18
+
+### Changed
+- Update to whelk 1.0.4
+
+### Fixed
+- Fix blank node subjects in [`report`] in [#767]
+
+## [1.7.1] - 2020-10-22
+
 ### Added
 - Add Whelk OWL reasoner in [#730]
 - Add [`validate`] command [#691]
 - Add `--errors <path>` option to [`template`] in [#713]
 - Add new formats to [`report`]: HTML, JSON, and XLSX [#699]
+- Add `--fail-on-violation <bool>` option to [`verify`] in [#738]
 
 ### Fixed
 - Handle empty [`template`] property charactersitics in [#719]
@@ -19,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix [`validate`] `--write-all true` in [#726]
 - Fix reported row number in [`validate`] error tables in [#727]
 - Fix equivalent class rendering for [`template`] in [#728]
+- Fix ontology IRI rendering for [`report`] in [#739]
 
 ## [1.7.0] - 2020-07-31
 
@@ -171,7 +183,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First official release of ROBOT!
 
-[Unreleased]: https://github.com/ontodev/robot/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/ontodev/robot/compare/v1.7.2...HEAD
+[1.7.2]: https://github.com/ontodev/robot/compare/v1.7.1...v1.7.2
+[1.7.1]: https://github.com/ontodev/robot/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/ontodev/robot/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/ontodev/robot/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/ontodev/robot/compare/v1.4.3...v1.5.0
@@ -205,6 +219,10 @@ First official release of ROBOT!
 [`template`]: http://robot.obolibrary.org/template
 [`validate`]: http://robot.obolibrary.org/validate
 
+[#767]: https://github.com/ontodev/robot/pull/767
+[#758]: https://github.com/ontodev/robot/pull/758
+[#739]: https://github.com/ontodev/robot/pull/739
+[#738]: https://github.com/ontodev/robot/pull/738
 [#728]: https://github.com/ontodev/robot/pull/728
 [#727]: https://github.com/ontodev/robot/pull/727
 [#726]: https://github.com/ontodev/robot/pull/726